Separated WsGateService from main
